shopping-basket: Connect signal to draw frame after inference
The signal that is emitted to draw the frame after inference is ran was
not connected to a slot and the GUI displayed the last frame before the
video feed was stopped. By adding a slot that draws the image to the
GUI, the frame that was used for inference is displayed with boxes from
the results.
Signed-off-by: Ayman Chaudhry <ayman.chaudhry.kc@renesas.com>