object-detection: Add object detection mode
The object detection mode draws boxes around detected objects and
displays the name and confidence of the object. This is controlled by a
start/stop button that changes between a continuous inference feed and a
live video feed. Inference and Total FPS are displayed on the top right
corner of the GUI. This mode allows the loading of a model and label
file.
Signed-off-by: Ayman Chaudhry <ayman.chaudhry.kc@renesas.com>