ui: Keep aspect ratio when displaying images and video
Changing aspect ratio to fit images and videos to the UI results in
a stretching effect. Keeping the aspect ratio of the source file allows
for a better presented image with additional whitespace on the UI.
Officially supported cameras do not need this change as we know the
expected shape of the output, whereas the user may choose an image or
video file of any shape.

Signed-off-by: Gareth Williams <gareth.williams.jx@renesas.com>