README.md: Update build instructions for XNNPack
The application supports XNNPack for the RZ/G2L and RZ/G2LC. However the
RZ/G2E and RZ/G2M do not include support due to the older verison of yocto
they are built with. The build instructions in the README.md file should
be updated to show the DEFINE option needed to include XNNPack support for
the relevant platforms.

Signed-off-by: Gareth Williams <gareth.williams.jx@renesas.com>