tfworker: Make Tensorflow-Lite output retrieval model generic
The class responsible for interacting with Tensorflow-Lite assumes set
outputs that are defined in the model. This means using models without
the same outputs is not possible. We should make the demo mode classes
responsible for interpreting output data whilst the tfworker should be
responsible for retrieving all information for the application to use.

Signed-off-by: Gareth Williams <gareth.williams.jx@renesas.com>