models: Ensure correct model selection when changing modes by GUI
Changing modes with the GUI can result in the model path for the
Shopping Basket mode not being set correctly. Restructuring when
this assignment is made fixes this. We should also save the model
path provided by the user for the Object Detection mode in case
the app is started in Shopping Basket mode.

Fixes: 820846c5c736

Signed-off-by: Gareth Williams <gareth.williams.jx@renesas.com>