tfworker: Set item stride based on demo mode
The item stride is currently being set by the last value in the output
tensor. Some models do not provide this so we should set the item stride
based on the demo mode being used. If the shopping basket or object
detection mode is being used, the item stride should be set as the
amount of outputs in the penultimate tensor. Pose Estimation should use
the last tensor.

Signed-off-by: Ayman Chaudhry <ayman.chaudhry.kc@renesas.com>