pose-estimation: Prevent disabling XNNPack when switching demo modes
When loading the BlazePose and HandPose models with TensorFlow Lite +
XNNPack delegate, the Inference Engine would switch back to TensorFlow
Lite and would have the XNNPack delegate option disabled. We should
correct this logic and check that the Inference Engine is TensorFlow
Lite + ArmNN delegate before defaulting to TensorFlow Lite.

Signed-off-by: Ayman Chaudhry <ayman.chaudhry.kc@renesas.com>