face-detection: Add Face Detection mode support
The Face Detection mode draws lines around the detected facial regions of
a person on the image frame and displays the face mesh on the 2D Point
Projection, which is located on the right side of the GUI. This is
controlled by a start/stop inference button that changes between a
continuous inference feed and a live video feed. Inference and Total FPS
are displayed on the top right corner of the GUI.

The model used in this demo mode requires the persons face to be close
to the camera to function accurately.

Model sourced from:
https://github.com/google/mediapipe/blob/master/mediapipe/modules/face_landmark/face_landmark.tflite

Signed-off-by: Ayman Chaudhry <ayman.chaudhry.kc@renesas.com>