face-detection: Add iris detection option
Support has been added for the iris landmark model for Face Detection
mode. Irises of a detected person will be drawn on the image frame when
the Detect Iris button has been selected. The pipeline of this option
uses the face detection to identify and crop the face, face landmark to
find the mesh of the face, and iris landmark is run twice to identify
the iris of both eyes. The inference times for each run of inference are
displayed on the top right corner of the GUI, with a diagram of the
pipeline displayed underneath.

Model sourced from:
https://github.com/google/mediapipe/blob/master/mediapipe/modules/iris_landmark/iris_landmark.tflite

Signed-off-by: Ayman Chaudhry <ayman.chaudhry.kc@renesas.com>