facedetection: Initialise face points during construction
The data structure used to track the facial points is allocated memory
with each frame during continuous inference however this memory is not
freed after use. Rather than using the delete operator at the end of the
function we can initialise the structure during object construction to
reduce processor overhead and allow the destructor to free the
memory instead.

Signed-off-by: Gareth Williams <gareth.williams.jx@renesas.com>