opencvworker: Refine opencv video capture code
Holding a VideoCapture device rather than a pointer to one allows us to
open the device at the same time it is initialised and simplify the
memory management of interacting with it. The format setting should also
be done after confirmation has been made that the VideoCapture device has
been opened sucessfully.

Signed-off-by: Gareth Williams <gareth.williams.jx@renesas.com>