audiocommand: Extract words from the mic
Up until now the application has been aquiring frames from
the microphone, and it's been sending samples at time T or
T - 0.5.

Switch to a trimming algorithm that determines where the word
is located in the input stream, and exctracts it centered
in the sample before sending it for inference.

The accuracy improvement is outstanding.

Signed-off-by: Fabrizio Castro <fabrizio.castro.jz@renesas.com>